+++ app/pb_kits/playbook/pb_circle_icon_button/docs/_circle_icon_button_link.jsx
@@ -1,0 +1,25 @@
+import React from 'react'
+import { CircleIconButton } from '../../'
+
+const CircleIconButtonLink = () => (
+  <div>
+    <CircleIconButton
+        icon="search"
+        link="https://google.com"
+        variant="primary"
+
+    />
+
+    <br />
+
+    <CircleIconButton
+        icon="window"
+        link="https://google.com"
+        newWindow
+        variant="secondary"
+    />
+
+  </div>
+)
+
+export default CircleIconButtonLink