+++ CHANGELOG.md
@@ -7,11 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+
 ### Added
 - Added legend prop to Bar Graph Kit and Line Graph Kit ([#818](https://github.com/powerhome/playbook/pull/818)@christinaatai)
 - Add React to LabelPill kit ([#813](https://github.com/powerhome/playbook/pull/813) @kellyeryan)
 - Updated Bar Graph Kit and Line Graph Kit with height prop ([#819](https://github.com/powerhome/playbook/pull/819)@christinaatai)
-
+- Added link & new_window props to circle_icon_button rails/react ()
 ### Fixed
 - Added aria, data, id props to Text Input kit on both React and Ruby sides.([#812](https://github.com/powerhome/playbook/pull/812)@kellyeryan)
 - Add aria to Ruby LabelPill kit ([#813](https://github.com/powerhome/playbook/pull/813) @kellyeryan)